pyproject.toml
[project]
name = "image-recoloring-GAN"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "graphviz>=0.20.3",
    "matplotlib>=3.10.1",
    "numpy>=2.2.3",
    "pillow>=11.1.0",
    "torch>=2.6.0",
    "torchview>=0.2.6",
]
